+++ src/main/java/insane96mcp/progressivebosses/module/dragon/feature/MinionFeature.java
@@ -143,8 +143,8 @@ public class MinionFeature extends Feature {
 		dragonTags.putInt(Strings.Tags.DRAGON_MINION_COOLDOWN, cooldown - 1);
 
 		float angle = world.random.nextFloat() * (float) Math.PI * 2f;
-		float x = (float) (Math.cos(angle) * (Mth.nextFloat(dragon.getRandom(), 16f, 45f)));
-		float z = (float) (Math.sin(angle) * (Mth.nextFloat(dragon.getRandom(), 16f, 45f)));
+		float x = (float) (Math.cos(angle) * (Mth.nextFloat(dragon.getRandom(), 16f, 40f)));
+		float z = (float) (Math.sin(angle) * (Mth.nextFloat(dragon.getRandom(), 16f, 40f)));
 		float y = world.getHeightmapPos(Heightmap.Types.MOTION_BLOCKING, new BlockPos(x, 255, z)).getY();
 		summonMinion(world, new Vec3(x, y, z), DifficultyHelper.getScalingDifficulty(dragon));
 	}