+++ src/main/java/insane96mcp/progressivebosses/module/dragon/feature/ResistancesFeature.java
@@ -23,7 +23,10 @@ public class ResistancesFeature extends Feature {
 
 	@Config(min = 0d, max = 1d)
 	@Label(name = "Melee Damage reduction while at the center", description = "Melee Damage reduction at max difficulty while the Ender Dragon is at the center.")
-	public static Double damageReductionWhenSitting = 0.40d;
+	public static Double damageReductionWhenSitting = 0.24d;
+	@Config(min = 0d, max = 1d)
+	@Label(name = "Melee Damage increase while not at the center", description = "Melee Damage is increased by this percentage while the Ender Dragon is not at the center.")
+	public static Double damageIncreaseWhileNotSitting = 0.24d;
 	@Config(min = 0d, max = 1d)
 	@Label(name = "Explosion Damage reduction", description = "Damage reduction when hit by explosions (firework rockets excluded).")
 	public static Double explosionDamageReduction = 0.667d;
@@ -39,6 +42,7 @@ public class ResistancesFeature extends Feature {
 			return;
 
 		meleeDamageReduction(event, dragon);
+		meleeDamageIncrease(event, dragon);
 		explosionDamageReduction(event, dragon);
 	}
 
@@ -49,6 +53,14 @@ public class ResistancesFeature extends Feature {
 			return;
 		if (sittingPhases.contains(dragon.getPhaseManager().getCurrentPhase().getPhase()) && event.getSource().getDirectEntity() instanceof Player) {
 			event.setAmount((event.getAmount() - (float) (event.getAmount() * (damageReductionWhenSitting * DifficultyHelper.getScalingDifficulty(dragon)))));
+		}
+	}
+
+	private static void meleeDamageIncrease(LivingDamageEvent event, EnderDragon dragon) {
+		if (damageIncreaseWhileNotSitting == 0d)
+			return;
+		if (!sittingPhases.contains(dragon.getPhaseManager().getCurrentPhase().getPhase()) && event.getSource().getDirectEntity() instanceof Player) {
+			event.setAmount((event.getAmount() + (float) (event.getAmount() * (damageIncreaseWhileNotSitting * DifficultyHelper.getScalingDifficulty(dragon)))));
 		}
 	}